scr/requirements.txt
# OCR 2.0 - Веб-приложение для обработки документов
# Основные зависимости для Flask веб-интерфейса

# Веб-фреймворк
Flask==2.3.3
Werkzeug==2.3.7

# OCR движки
paddlepaddle==2.5.1
paddleocr==2.7.0.3
pytesseract==0.3.10
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2

# Обработка изображений
Pillow==10.0.0
opencv-python==4.8.0.76
pdf2image==1.16.3
PyMuPDF==1.23.8

# Анализ и постобработка данных
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# NLP и текстовый анализ
nltk==3.8.1
python-Levenshtein==0.21.1
editdistance==0.6.2

# LLM интеграция (опционально)
openai==0.28.1
requests==2.31.0

# Метрики и валидация
jiwer==3.0.3

# Утилиты
tqdm==4.66.1
pathlib2==2.3.7
python-dateutil==2.8.2

# Веб-интерфейс (альтернативный Streamlit)
streamlit==1.27.0
plotly==5.15.0

# Для разработки (опционально)
pytest==7.4.2
black==23.7.0
flake8==6.0.0

# Системные зависимости (документация)
# sudo apt-get install tesseract-ocr tesseract-ocr-rus tesseract-ocr-eng
# brew install tesseract tesseract-lang  # для macOS